djangoRestProject/requirements.txt

aniso8601==8.0.0
beautifulsoup4
certifi==2020.6.20
chardet==3.0.2
cmake
click==7
dlib
django-cors-headers
djangorestframework
folium==0.12.1
Flask-RESTful==0.3.8
Flask==1.1.2
googlemaps==4.4.5
gunicorn==20.0.4
html5lib
icecream==2.1.1
idna==2.10
imbalanced-learn
itsdangerous==1.1.0
JPype1
jsonify==0.5
jupyter==1
konlpy==0.5.2
lxml
matplotlib
MarkupSafe==1.1.1
mccabe
mysql-connector
mysqlclient
nltk
numpy
opencv-python~=4.6.0.66
pandas
pandas-datareader==0.9.0
pep8==1.7.1
plotly==5
python-dateutil==2.8.1
pytz==2020.5
requests==2.24.0
seaborn
selenium==3.141.0
six==1.15.0
scikit-learn
SQLAlchemy==1.4.39
tensorflow-gpu==2.6.0
tensorflow==2.7.0
tensorflow_datasets
torch
torchvision
tweepy==3.10.0
tqdm
urllib3==1.25.10
Werkzeug==1
wincertstore==0.2
wordcloud
yfinance == 0.2.3
prophet == 1.1.1

apache_beam